# Optimize log hot path