Followup to #4891.

Reduces allocations on happy path by avoiding creation of extra `InstrumentationScopeInfo` / `InstrumentationScopeInfoBuilder` instances that aren't strictly needed. Does this by adjusting changing `ComponentRegistry#get(InstrumentationScopeInfo)` to `ComponentRegistry#get(String name, @Nullable String version, @Nullable String schemaUrl, Attributes attributes)`, and only building `InstrumentationScopeInfo` the first time we see a component with a particular identity. 

Benchmark before:
```
Benchmark                                                  Mode  Cnt     Score     Error   Units
LogsBenchmarks.emitSimpleLog                               avgt   10    35.035 ±   0.732   ns/op
LogsBenchmarks.emitSimpleLog:·gc.alloc.rate                avgt   10  2173.006 ±  43.649  MB/sec
LogsBenchmarks.emitSimpleLog:·gc.alloc.rate.norm           avgt   10   120.000 ±   0.001    B/op
LogsBenchmarks.emitSimpleLog:·gc.churn.G1_Eden_Space       avgt   10  2191.457 ± 429.679  MB/sec
LogsBenchmarks.emitSimpleLog:·gc.churn.G1_Eden_Space.norm  avgt   10   120.983 ±  23.200    B/op
LogsBenchmarks.emitSimpleLog:·gc.count                     avgt   10    37.000            counts
LogsBenchmarks.emitSimpleLog:·gc.time                      avgt   10    30.000                ms
```

Benchmark after:
```
Benchmark                                                  Mode  Cnt     Score     Error   Units
LogsBenchmarks.emitSimpleLog                               avgt   10    29.155 ±   1.023   ns/op
LogsBenchmarks.emitSimpleLog:·gc.alloc.rate                avgt   10  1220.007 ±  36.921  MB/sec
LogsBenchmarks.emitSimpleLog:·gc.alloc.rate.norm           avgt   10    56.048 ±   0.231    B/op
LogsBenchmarks.emitSimpleLog:·gc.churn.G1_Eden_Space       avgt   10  1241.334 ± 191.229  MB/sec
LogsBenchmarks.emitSimpleLog:·gc.churn.G1_Eden_Space.norm  avgt   10    57.041 ±   8.784    B/op
LogsBenchmarks.emitSimpleLog:·gc.count                     avgt   10    31.000            counts
LogsBenchmarks.emitSimpleLog:·gc.time                      avgt   10    27.000                ms
```

Notice the significant reduction in allocation rate.