@jkwatson 2022-11-08
> > Shouldn't people be initializing their loggers/tracers/etc once, rather than fetching them every time?
> 
> With tracers and meters, yes. But log appender implementations use the logger name as the name of the logger, so they have to get a logger for each log record they process. Here's an [example](https://github.com/open-telemetry/opentelemetry-java-instrumentation/blob/main/instrumentation/log4j/log4j-appender-2.17/library/src/main/java/io/opentelemetry/instrumentation/log4j/appender/v2_17/OpenTelemetryAppender.java#L143-L144) from the log4j appender. The logback and jul implementations work the same way. The appenders _could_ implement their own cache of loggers by maintaining a map of logger name to logger, but if that's a pattern all implementations will have to do we should probably solve it in the SDK.

ah, I see. for the appender case, that makes sense. thanks!